requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.6.2=pyhd8ed1ab_0
arrow=1.2.3=pyhd8ed1ab_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.12.2=pyha770c72_0
black=23.3.0=pypi_0
blessed=1.19.1=pyhe4f9e05_2
brotlipy=0.7.0=py39hb9d737c_1004
ca-certificates=2023.5.7=hbcca054_0
cachecontrol=0.12.11=pyhd8ed1ab_1
cachy=0.3.0=pyhd8ed1ab_1
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.14.6=py39he32792d_0
charset-normalizer=3.1.0=pyhd8ed1ab_0
cleo=0.8.1=pyhd8ed1ab_2
click=8.1.3=unix_pyhd8ed1ab_2
clikit=0.6.2=pyhd8ed1ab_2
cmake=3.26.3=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
crashtest=0.3.1=pyhd8ed1ab_0
croniter=1.3.14=pyhd8ed1ab_0
cryptography=37.0.2=py39hd97740a_0
dateutils=0.6.12=py_0
dbus=1.13.0=h4e0c4b3_1000
deepdiff=6.2.2=pyhd8ed1ab_0
distlib=0.3.6=pyhd8ed1ab_0
efficientnet-pytorch=0.7.1=pypi_0
expat=2.4.8=h27087fc_0
fastapi=0.88.0=pyhd8ed1ab_0
filelock=3.12.0=pyhd8ed1ab_0
fsspec=2023.5.0=pyh1a96a4e_0
future=0.18.3=pyhd8ed1ab_0
h11=0.14.0=pyhd8ed1ab_0
html5lib=1.1=pyh9f0ad1d_0
huggingface-hub=0.14.1=pypi_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
inquirer=3.1.3=pyhd8ed1ab_0
itsdangerous=2.1.2=pyhd8ed1ab_0
jaraco.classes=3.2.3=pyhd8ed1ab_0
jeepney=0.8.0=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
keyring=23.13.1=py39hf3d152e_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=15_linux64_openblas
libopenblas=0.3.20=pthreads_h78a6416_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
lightning=2.0.2=pyhd8ed1ab_0
lightning-cloud=0.5.35=pyhd8ed1ab_0
lightning-utilities=0.8.0=pyhd8ed1ab_0
lit=16.0.5=pypi_0
lockfile=0.12.2=py_1
markdown-it-py=2.2.0=pyhd8ed1ab_0
markupsafe=2.1.1=py39hb9d737c_1
mdurl=0.1.0=pyhd8ed1ab_0
more-itertools=9.1.0=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
msgpack-python=1.0.3=py39hf939315_1
munch=3.0.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
ninja=1.11.0=h924138e_0
numpy=1.22.3=py39hc58783e_2
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1o=h166bdaf_0
ordered-set=4.1.0=pyhd8ed1ab_0
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
pastel=0.2.1=pyhd8ed1ab_0
pathspec=0.11.1=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pillow=9.5.0=pypi_0
pip=23.0.1=py39h06a4308_0
pkginfo=1.9.6=pyhd8ed1ab_0
platformdirs=3.5.1=pyhd8ed1ab_0
poetry=1.1.15=py39hf3d152e_0
poetry-core=1.0.8=py39hf3d152e_1
pretrainedmodels=0.7.4=pypi_0
psutil=5.9.1=py39hb9d737c_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.9.1=py39hb9d737c_0
pygments=2.15.1=pyhd8ed1ab_0
pyjwt=2.7.0=pyhd8ed1ab_0
pylev=1.4.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.0=hdb3f193_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python-editor=1.0.4=py_0
python-multipart=0.0.6=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch-lightning=2.0.2=pyhd8ed1ab_0
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_4
readchar=4.0.5=pyhd8ed1ab_0
readline=8.2=h5eee18b_0
requests=2.29.0=pyhd8ed1ab_0
requests-toolbelt=0.9.1=py_0
rich=13.3.5=pyhd8ed1ab_0
secretstorage=3.3.3=py39hf3d152e_1
segmentation-models-pytorch=0.3.2=pypi_0
setuptools=66.0.0=py39h06a4308_0
shellingham=1.5.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.41.2=h5eee18b_0
starlette=0.22.0=pyhd8ed1ab_0
starsessions=1.3.0=pyhd8ed1ab_0
sympy=1.12=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
tomlkit=0.11.8=pyha770c72_0
torch=2.0.1=pypi_0
torchmetrics=0.11.4=pyhd8ed1ab_0
torchvision=0.15.2=pypi_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pyhd8ed1ab_0
triton=2.0.0=pypi_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=1.26.15=pyhd8ed1ab_0
uvicorn=0.22.0=py39hf3d152e_0
virtualenv=20.23.0=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py_1
websocket-client=1.5.1=pyhd8ed1ab_0
websockets=10.3=py39hb9d737c_0
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
